Amend/append site description
As of writing, the description shown on search engines is most off-putting, as
it quotes the rule against CSAM, likely because it's one of the first mentions
of "CollabVM".

This commit aims to try and fix that by explicitly mentioning the site name in
the meta tags, and by adding a site description at the beginning of the rules.

All other translations I imagine would have to be updated to reflect these, as
the same issue would perhaps show if you searched CollabVM in another country.

(Side note, why is en-US duplicated in two files?)